attached_assets/requirements_1756037623259.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
pandas==2.2.2
pytest==7.4.2
numpy>=1.26.0
